requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.1
langchain-groq>=0.1.0
langchain-community>=0.0.20
pydantic>=2.5.0
streamlit>=1.30.0
python-dotenv>=1.0.0
plotly>=5.18.0
huggingface-hub>=0.20.0
bcrypt>=4.1.0
sqlalchemy>=2.0.0
